backend/pyproject.toml
[tool.poetry]
name = "quikbite-backend"
version = "0.1.0"
description = "Backend for QuikBite CMS"
authors = ["Kevin Nadar <kevinxaviernadar@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
SQLAlchemy = "^2.0.25"
python-jose = "^3.3.0"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.1.2"
pyotp = "^2.9.0"
pandas = "^2.1.4"
xlrd = "^2.0.1"
pyqt5 = "^5.15.10"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
fastapi-mail = "^1.4.1"
uvicorn = "^0.27.0.post1"
requests = "^2.31.0"
httpx = "^0.26.0"
pytest = "^8.1.1"
setuptools = "^69.2.0"
razorpay = "^1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
